Rename 'api' folder to 'utils'
