+++ browser-extension/src/components/HistoryPage.tsx
@@ -1,6 +1,6 @@
 import { CircularProgress, Typography } from "@material-ui/core";
 import { FC, useContext, useEffect, useState } from "react";
-import client from "../api/client";
+import client from "../utils/client";
 import { History } from "../types";
 import AppContext from "./AppContext";
