+++ browser-extension/src/utils/client.ts
