+++ browser-extension/src/utils/persistentStorage.ts
